fix:transaction modul

+++ app/Http/Controllers/BiayaController.php
@@ -6,8 +6,7 @@ use App\Http\Requests\StoreBiayaRequest;
 use App\Http\Requests\UpdateBiayaRequest;
 use Illuminate\Http\Request;
 use App\Models\Biaya as Model;
-
-
+use App\Models\Biaya;
 
 class BiayaController extends Controller
 {
@@ -22,10 +21,10 @@ class BiayaController extends Controller
     public function index(Request $request)
     {
         if ($request->filled('q')) {
-            $models = Model::with('user')->search($request->q)->paginate(50);
+            $models = Model::with('user')->whereNull('parent_id')->search($request->q)->paginate(50);
         } else {
             /** eager loading pake (with untuk relasi) */
-            $models = Model::with('user')->latest()->paginate(50);
+            $models = Model::with('user')->whereNull('parent_id')->latest()->paginate(50);
         }
 
         $data = [
@@ -39,14 +38,20 @@ class BiayaController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        $biaya = new Model();
+        if ($request->filled('parent_id')) {
+            $biaya = Biaya::with('children')->findOrFail($request->parent_id);
+        }
+        // dd($biaya);
         $data = [
-            'model'     => new Model(),
-            'method'    => 'POST',
-            'route'     => $this->routePrefix . '.store',
-            'button'    => 'SIMPAN',
-            'title'     => 'Tambah Biaya'
+            'parentData'    => $biaya,
+            'model'         => new Model(),
+            'method'        => 'POST',
+            'route'         => $this->routePrefix . '.store',
+            'button'        => 'SIMPAN',
+            'title'         => 'Tambah Biaya'
         ];
 
         return view('operator.' . $this->viewCreate, $data);
@@ -58,7 +63,11 @@ class BiayaController extends Controller
     public function store(StoreBiayaRequest $request)
     {
         Model::create($request->validated());
-        return  redirect()->route($this->routePrefix . '.index')->with('success', 'Data Biaya berhasil di simpan.');
+        if ($request->filled('parent_id')) {
+            return back()->with('success', 'Data Biaya berhasil di simpan.');
+        } else {
+            return  redirect()->route($this->routePrefix . '.index')->with('success', 'Data Biaya berhasil di simpan.');
+        }
     }
 
     /**
@@ -107,7 +116,21 @@ class BiayaController extends Controller
     public function destroy(string $id)
     {
         $model = Model::findOrFail($id);
+        $children = $model->children->count();
+
+        if ($children > 0) {
+            return back()->with('error', 'Data tidak bisa dihapus karena masih memiliki item biaya. Hapus item biaya, terlebih dahulu!');
+        }
+
         $model->delete();
         return redirect()->route($this->routePrefix . '.index')->with('success', 'Data berhasil di hapus.');
+    }
+
+    public function deleteItem(string $id)
+    {
+        $model = Model::findOrFail($id);
+        $parentId = $model->parent_id;
+        $model->delete();
+        return back()->with('success', 'Data Item berhasil di hapus.');
     }
 }
